Refactor channel_reestablish creation
To remove duplication between offline and syncing states.
